+++ gdscript-mode.el
@@ -31,6 +31,7 @@
 ;;; Code:
 
 (require 'gdscript-customization)
+(require 'gdscript-documentation)
 (require 'gdscript-syntax)
 (require 'gdscript-indent-and-nav)
 (require 'gdscript-imenu)
@@ -71,6 +72,8 @@
                             (define-key map "\C-c\C-r\C-q" 'gdscript-godot-run-current-scene-debug)
                             (define-key map "\C-c\C-r\C-e" 'gdscript-godot-edit-current-scene)
                             (define-key map "\C-c\C-r\C-x" 'gdscript-godot-run-current-script)
+                            (define-key map "\C-c\C-r\C-a" 'gdscript-documentation-api)
+                            (define-key map "\C-c\C-r\C-o" 'gdscript-documentation-open)
                             map)
   "Keymap for `gdscript-mode'.")